rename state files to 'simulator'
