+++ intelligence/user-simulation/src/api_simulator.js
