+++ intelligence/user-simulation/src/dummy_simulator.js
