+++ intelligence/user-simulation/src/night_simulator.js
